+++ plugins/card/build.js
@@ -3,9 +3,12 @@ module.exports = {
     'my-card': '../card/card'
   },
   handler (file) {
-    // 删去原来的 card 标签
-    if (file.basename === 'node.wxml' || file.basename === 'node.vue') {
-      file.contents = Buffer.from(file.contents.toString().replace(/<card[\s\S]+?>/, ''))
+    if (file.isBuffer()) {
+      let content = file.contents.toString()
+      if (file.path.includes('parser.js')) {
+        content = content.replace(/trustTags\s*:\s*makeMap\('/, "trustTags: makeMap('card,").replace(/voidTags\s*:\s*makeMap\('/, "voidTags: makeMap('card,")
+      }
+      file.contents = Buffer.from(content)
     }
   }
 }