+++ plugins/editable/uni-app/build.js
@@ -281,14 +281,14 @@ module.exports = {
      */
     mediaTap (e, index) {
       if (this.opts[5]) {
-        const i = index || e.target.dataset.i
+        const i = e.target.dataset.i || index
         const node = this.childs[i]
         const items = this.root._getItem(node)
         this.root._maskTap()
         this.root._edit = this
         this.i = i
         this.root._tooltip({
-          top: (index != undefined ? e.currentTarget.offsetTop: e.target.offsetTop) - 30,
+          top: e.currentTarget.offsetTop - 30,
           items,
           success: tapIndex => {
             switch (items[tapIndex]) {
@@ -316,37 +316,6 @@ module.exports = {
                 uni.showToast({
                   title: '成功'
                 })
-                break
-            }
-          }
-        })
-        // 避免上层出现点击态
-        this.root._lock = true
-        setTimeout(() => {
-          this.root._lock = false
-        }, 50)
-      }
-    },
-    
-    /**
-     * @description 自定义Node被点击
-     * @param {Event} e
-     */
-    nodeClick (e, index) {
-      if (this.opts[5]) {
-        const i = index || e.currentTarget.dataset.i
-        const node = this.childs[i]
-        const items = this.root._getItem(node);
-        this.root._maskTap()
-        this.root._edit = this
-        this.i = i
-        this.root._tooltip({
-          top: e.currentTarget.offsetTop - 30,
-          items,
-          success: tapIndex => {
-            switch (items[tapIndex]) {
-              case '删除':
-                this.remove(i)
                 break
             }
           }
@@ -553,7 +522,7 @@ module.exports = {
             .replace('<video', '<video :show-center-play-btn="!opts[5]" @tap="mediaTap"')
             .replace('<audio ', '<audio @tap="mediaTap" ')
             .replace('<my-audio ', '<my-audio @onClick="mediaTap($event, i)" ')
-            .replace('card ', 'card @onClick="nodeClick($event, i)" ')
+            .replace('card ', 'card @click="mediaTap($event, i)" ')
             .replace('<script>',
               `<script>
 import Parser from '../parser'
@@ -578,7 +547,7 @@ function getTop(e) {
 }`)
             // 周期处理
             .replace(/beforeDestroy\s*\(\)\s*{/, `beforeDestroy () {
-  if (this.root._edit === this) {
+  if (this.root && this.root._edit === this) {
     this.root._edit = undefined
   }`)
             // 记录图片宽度