+++ test/test.js
@@ -62,7 +62,8 @@ console.log('11')
     <a>xxx</a>
   </td>
 </table>
-<table width="100%" border="1">
+<table width="100%" border="1" style="border-collapse: collapse;">
+  <colgroup><col style="width: 40%;"><col style="width: 60%;"></colgroup>
   <tr>
     <th style="vertical-align: bottom;" width="20%">标题1</th>
     <th width="80%">标题2</th>
@@ -88,6 +89,8 @@ console.log('11')
   <pre>var i = 0</pre>
 </div>
 <!-- 测试不同情况中的图片处理 -->
+<img src="xxx" style="width:100px;height:100px;object-fit:contain">
+<img src="xxx" style="width:100px;object-fit:cover;height:100px;">
 <a data-test="test">
   <img src="//xxx.jpg">
 </a>
@@ -105,6 +108,7 @@ console.log('11')
 <img src="yyy.webp" style="width:1000px" ignore>
 <svg />
 <svg viewbox="0 0 1 1"><text>123</text><svg></svg></svg>
+<svg><foreignobject><div>123</div></foreignobject></svg>
 <div class="ql-align-center" style="background-image:url(&quot;/xxx.jpg?a=2&amp;b=3&quot;)"></div>
 <![CDATA[<]]>
 <!-- 测试 flex 布局、未闭合标签、data- 属性处理 -->